requirements.txt
amqp==2.5.2
billiard==3.6.1.0
Brotli==1.0.7
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.2.0
django-extensions==2.2.5
django-storages==1.8
djangorestframework==3.10.3
dnspython==1.16.0
dropbox==9.4.0
eventlet==0.25.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.0
idna==2.8
importlib-metadata==1.3.0
kombu==4.6.7
monotonic==1.5
more-itertools==8.0.2
psycopg2==2.8.4
pycparser==2.19
pytz==2019.2
redis==3.3.11
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
whitenoise==4.1.4
zipp==0.6.0
